(build) Update dependencies and fix tests
Weird gotcha is that Jooby seems to have changed the generated service suffix to add Htmx for some reason.

+++ code/functions/language-processing/test/nu/marginalia/keyword/DocumentKeywordExtractorTest.java
@@ -47,7 +47,7 @@ class DocumentKeywordExtractorTest {
 
         keywords.getWordToMeta().forEach((k, v) -> {
             if (k.contains("_")) {
-                System.out.println(k + " " + WordFlags.decode((byte) v.longValue()));
+                System.out.println(k + " " + WordFlags.decode((byte) v));
             }
         });
     }
